refactor: move benchmark suite to a separate crate
This removes the (dev-) dependency on GSL wrapper in gomez itself. The
benchmarks were also ported to use divan instead of criterion.
